일기관련 엔터티의 id를 long형으로 변경

+++ src/main/java/com/fooddiary/api/dto/response/image/ImageResponseDTO.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ImageResponseDTO {
-    private int imageId;
+    private long imageId;
     private byte[] bytes;
 }
